requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
Pillow==10.1.0
httpx==0.25.1
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
numpy==1.26.2
